integration_service/requirements.txt
flask-cors>=4.0.0
flask==2.3.2
celery==5.3.1  
redis==4.5.1  
requests==2.31.0
GitPython==3.1.42
PyGithub==1.58.0       # GitHub API
python-gitlab>=3.8.0  # GitLab API
bitbucket-api==0.5.0   # Bitbucket API
jira==3.5.0            # Jira API